+++ projects/sobal/index.js
@@ -1,1 +1,14 @@
-module.exports = require('./onchain')
+const { onChainTvl } = require('../helper/balancer')
+
+const V2_ADDRESS = '0x7122e35ceC2eED4A989D9b0A71998534A203972C'; // Vault
+
+const config = {
+  neon_evm: { fromBlock: 206166057, },
+}
+
+Object.keys(config).forEach(chain => {
+  const { fromBlock } = config[chain]
+  module.exports[chain] = {
+    tvl: onChainTvl(V2_ADDRESS, fromBlock)
+  }
+})
